requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
fastapi==0.134.0
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
kombu==5.6.2
langchain-core==1.2.16
langchain-groq==1.1.2
mako==1.3.10
markupsafe==3.0.3
packaging==26.0
prompt-toolkit==3.0.52
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart
pyyaml==6.0.3
redis==7.2.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.47
starlette==0.52.1
structlog==25.5.0
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn[standard]==0.41.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
google-cloud-storage
google-auth
